requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7.3
django-autoslug==1.7.2
django-cities-light==3.1.2
django-classy-tags==0.6.1
django-cms==3.0.10
django-guardian==1.2.5
django-mptt==0.6.1
django-polls==0.2
django-reversion==1.8.5
django-sekizai==0.8.1
django-sslify==0.2.7
django-toolbelt==0.0.1
docutils==0.13
gunicorn==19.3.0
html5lib==0.999
Pillow==2.7.0
psycopg2==2.5.4
pytz==2014.10
requests==2.5.1
six==1.9.0
South==1.0.2
static3==0.6.1
Unidecode==0.4.17
